% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{abind}
\alias{abind}
\title{Combine multi-dimensional array-like objects.}
\usage{
abind(..., along)
}
\arguments{
\item{...}{Any number of array-like objects. The dimensions of all the
array-like objects must match, except on one dimension (specified by
\code{along}).}

\item{along}{The dimension along which to bind the array-like objects.}
}
\value{
Specific methods will typically return an object of the same class
as the input objects.
}
\description{
Combine multi-dimensional array-like objects. abind is a generalization of
cbind and rbind inspired by \code{abind::\link{abind}}. Takes a sequence of
array-like objects and produces a single object of the same class with the
same or higher dimension.

NOTE: This man page is for the \code{abind} \emph{S4 generic functions}
defined in the \pkg{DSArray} package. See \code{?abind::\link{abind}} for
the default methods (defined in the \pkg{abind} package) and which is the
inspiration for this S4 generic.
}
\examples{
abind  # note the dispatch on the '...' arg only
showMethods("abind")
selectMethod("abind", "DSArray")  # the DSArray method

}
\seealso{
\itemize{
   \item \code{abind::\link[abind]{abind}} for the inspiration for this
         generic.
 }
}

