% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{densify}
\alias{densify}
\title{Densify a DSArray object}
\usage{
densify(x)
}
\arguments{
\item{x}{A \link{DSArray} object.}
}
\value{
A 3-dimensional \link[base]{array}.
}
\description{
Densifying a DSArray object coerces it to the equivalent 3-dimensional
\link[base]{array}.
}
\details{
If \code{x} is a DSArray object then \code{densify(x)}
is equivalent to \code{as(x, "array")}.
}

