% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRMatrix.R
\docType{package}
\name{DRMatrix-package}
\alias{DRMatrix-package}
\title{DRMatrix: compressed storage of numeric matrices with many duplicate rows.}
\description{
DRMatrix provides efficient storage of numeric matrices with many duplicate
rows via the DRMatrix (\emph{D}uplicate \emph{R}ow \emph{Matrix}) S4 class.
A basic matrix-like API is provided for instantiating, subsetting, and
combining DRMatrix objects.
}
\details{
To learn more about DRMatrix, start with the vignette:
\code{browseVignettes(package = "DRMatrix")}
}

