% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRMatrix-class.R
\docType{methods}
\name{DRMatrix,matrix-method}
\alias{DRMatrix,array-method}
\alias{DRMatrix,list-method}
\alias{DRMatrix,matrix-method}
\alias{DRMatrix,missing-method}
\title{DRMatrix constructor}
\usage{
\S4method{DRMatrix}{matrix}(x, dimnames = NULL)

\S4method{DRMatrix}{missing}(x, dimnames = NULL)

\S4method{DRMatrix}{list}(x, dimnames = NULL)

\S4method{DRMatrix}{array}(x, MARGIN, dimnames = NULL)
}
\arguments{
\item{x}{A \link[base]{matrix}, \link[base]{list} of \link[base]{matrix}
objects with identical dimensions, or a 3-dimensional \link[base]{array}.}

\item{dimnames}{A \link[base]{dimnames} attribute for the DRMatrix: NULL or
a list of length 3. An empty list is treated as \code{NULL}, a list of
length one as row names, and a list of length two as row names and column
names. \strong{TODO: Can the list be named, and are the list names used as
names for the dimensions (see ?matrix)}. If \code{NULL}, the \code{dimnames}
are constructed from \code{x}; see Dimnames.}

\item{MARGIN}{An integer given the dimension number that distinguishes
samples; see Examples.}
}
\value{
A \link[=DRMatrix-class]{DRMatrix} object.
}
\description{
Construct a \link[=DRMatrix-class]{DRMatrix} from a \link[base]{matrix}, a
\link[base]{list} of \link[base]{matrix} objects with identical dimensions,
or a 3-dimensional \link[base]{array}.
}
\section{Dimnames}{

\strong{TODO}
}

