% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtilities.R
\name{dsa_identical_to_array}
\alias{dsa_identical_to_array}
\title{Check whether DSArray representation is identical to array representation}
\usage{
dsa_identical_to_array(dsarray, array)
}
\arguments{
\item{dsarray}{A DSArray}

\item{array}{A base::array}
}
\value{
\code{TRUE} if identical, \code{FALSE} otherwise
}
\description{
Check whether DSArray representation is identical to array representation
}
\keyword{internal}

