% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSArray.R
\docType{package}
\name{DSArray-package}
\alias{DSArray-package}
\title{DSArray: compressed storage of 3-dimensional arrays with many duplicate
slices.}
\description{
DSArray provides efficient storage of 3-dimensional arrays with many
duplicate slices via the DSArray (\emph{D}uplicate \emph{S}lice
\emph{Array}) S4 class. A basic array-like API is provided for
instantiating, subsetting, and combining DSArray objects.
}
\details{
To learn more about DSArray, start with the vignette:
\code{browseVignettes(package = "DSArray")}
}

