% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nslice}
\alias{nslice}
\title{The number of \emph{slices} of a \link[=DSArray-class]{DSArray}}
\usage{
nslice(x)
}
\arguments{
\item{x}{A \link[=DSArray-class]{DSArray} object.}
}
\value{
An \link[base]{integer} of length 1.
}
\description{
Analogous to \code{\link[base]{nrow}} and \code{\link[base]{nrow}},
\code{nslice} returns the number of \emph{slices} in a
\link[=DSArray-class]{DSArray} object. If \code{x} is a
\link[=DSArray-class]{DSArray} object then \code{nslice(x)} is equivalent
to \code{dim(x)[3]}.
}

