% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSArray-class.R
\docType{methods}
\name{DSArray,matrix-method}
\alias{DSArray,array-method}
\alias{DSArray,list-method}
\alias{DSArray,matrix-method}
\alias{DSArray,missing-method}
\title{DSArray constructor}
\usage{
\S4method{DSArray}{matrix}(x, dimnames = NULL)

\S4method{DSArray}{missing}(x, dimnames = NULL)

\S4method{DSArray}{list}(x, dimnames = NULL)

\S4method{DSArray}{array}(x, MARGIN = 2L, dimnames = NULL)
}
\arguments{
\item{x}{A 3-dimensional \link[base]{array}, a \link[base]{list} of
\link[base]{matrix} objects with identical dimensions, or a
\link[base]{matrix}.}

\item{dimnames}{A \link[base]{dimnames} attribute for the DSArray: NULL or
a list of length 3. An empty list is treated as \code{NULL}, a list of
length one as row names, and a list of length two as row names and column
names. \strong{TODO: Can the list be named, and are the list names used as
names for the dimensions (see ?array)}. If \code{NULL}, the \code{dimnames}
are constructed from \code{x}; see Dimnames.}

\item{MARGIN}{An integer given the dimension number that indexes samples;
see Examples. The default, \code{MARGIN = 2}, corresponds to columns of
\code{x} indexing samples.}
}
\value{
A \link[=DSArray-class]{DSArray} object.
}
\description{
Construct a \link[=DSArray-class]{DSArray} from a 3-dimensional
\link[base]{array}, a \link[base]{matrix}, or a \link[base]{list} of
\link[base]{matrix} objects with identical dimensions.
}
\section{Dimnames}{

\strong{TODO: Document how dimnames are constructed.}
}

